# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A production-ready Job Portal backend with role-based access control.",
        "title": "Job Portal API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@jobportal.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/register": {
            "post": {
                "description": "Create a new account with name, email, password and role",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Auth"],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Register payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": { "$ref": "#/definitions/RegisterInput" }
                    }
                ],
                "responses": {
                    "201": { "description": "Created", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "409": { "description": "Conflict", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticate with email and password, returns a JWT token",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Auth"],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": { "$ref": "#/definitions/LoginInput" }
                    }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "401": { "description": "Unauthorized", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            }
        },
...
        "/jobs/{id}/apply": {
            "post": {
                "security": [{ "BearerAuth": [] }],
                "description": "Candidate submits an application for a job",
                "produces": ["application/json"],
                "tags": ["Applications"],
                "summary": "Apply for a job",
                "parameters": [
                    { "type": "string", "description": "Job UUID", "name": "id", "in": "path", "required": true }
                ],
                "responses": {
                    "201": { "description": "Created", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "409": { "description": "Conflict", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            }
        },
        "/applications": {
            "get": {
                "security": [{ "BearerAuth": [] }],
                "description": "Candidate views their own job applications",
                "produces": ["application/json"],
                "tags": ["Applications"],
                "summary": "My applications",
                "parameters": [
                    { "type": "integer", "default": 1, "description": "Page number", "name": "page", "in": "query" },
                    { "type": "integer", "default": 10, "description": "Items per page", "name": "page_size", "in": "query" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "401": { "description": "Unauthorized", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            }
        },
        "/applications/job/{id}": {
            "get": {
                "security": [{ "BearerAuth": [] }],
                "description": "Recruiter views all applicants for one of their jobs",
                "produces": ["application/json"],
                "tags": ["Applications"],
                "summary": "Applicants for a job",
                "parameters": [
                    { "type": "string", "description": "Job UUID", "name": "id", "in": "path", "required": true },
                    { "type": "integer", "default": 1, "description": "Page number", "name": "page", "in": "query" },
                    { "type": "integer", "default": 10, "description": "Items per page", "name": "page_size", "in": "query" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            }
        },
        "/applications/{id}/status": {
            "patch": {
                "security": [{ "BearerAuth": [] }],
                "description": "Recruiter updates the status of an application",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Applications"],
                "summary": "Update application status",
                "parameters": [
                    { "type": "string", "description": "Application UUID", "name": "id", "in": "path", "required": true },
                    {
                        "description": "Status payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": { "$ref": "#/definitions/UpdateStatusInput" }
                    }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            }
        },
...
        "/admin/applications": {
            "get": {
                "security": [{ "BearerAuth": [] }],
                "description": "Admin retrieves all applications across all jobs",
                "produces": ["application/json"],
                "tags": ["Admin"],
                "summary": "[Admin] List all applications",
                "parameters": [
                    { "type": "integer", "default": 1, "description": "Page number", "name": "page", "in": "query" },
                    { "type": "integer", "default": 10, "description": "Items per page", "name": "page_size", "in": "query" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "401": { "description": "Unauthorized", "schema": { "$ref": "#/definitions/APIResponse" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/APIResponse" } }
                }
            }
        }
    },
    "definitions": {
        "APIResponse": {
            "type": "object",
            "properties": {
                "success": { "type": "boolean" },
                "message": { "type": "string" },
                "data": {},
                "error": {}
            }
        },
        "RegisterInput": {
            "type": "object",
            "required": ["name", "email", "password", "role"],
            "properties": {
                "name":     { "type": "string", "example": "Alice Smith" },
                "email":    { "type": "string", "example": "alice@example.com" },
                "password": { "type": "string", "example": "secret123" },
                "role":     { "type": "string", "enum": ["ADMIN", "RECRUITER", "CANDIDATE"], "example": "CANDIDATE" }
            }
        },
        "LoginInput": {
            "type": "object",
            "required": ["email", "password"],
            "properties": {
                "email":    { "type": "string", "example": "alice@example.com" },
                "password": { "type": "string", "example": "secret123" }
            }
        },
...
        "UpdateStatusInput": {
            "type": "object",
            "required": ["status"],
            "properties": {
                "status": { "type": "string", "enum": ["APPLIED", "REVIEWED", "REJECTED", "ACCEPTED"], "example": "ACCEPTED" }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter your JWT token as: Bearer \u003ctoken\u003e"
        }
    }
}